requirements_train.txt
# Trading Project Requirements - FOR TRAINING (e.g., Google Colab)
# Core dependencies
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.1.0
lightgbm>=3.3.0
xgboost>=1.6.0
matplotlib>=3.5.0
seaborn>=0.11.0
requests>=2.28.0
joblib>=1.1.0

# Data sources
yfinance>=0.2.0
fredapi>=0.5.0
pandas-datareader>=0.10.0

# Text processing
nltk>=3.8
textblob>=0.19.0
transformers>=4.30.0
torch>=2.0.0
tensorflow>=2.13.0

# Web scraping
beautifulsoup4>=4.12.0
feedparser>=6.0.0
lxml>=4.9.0

# Technical analysis
ta>=0.11.0

# Database
sqlalchemy>=2.0.0

# Configuration
python-dotenv>=1.0.0
pyyaml>=6.0

# Utilities
tqdm>=4.65.0
psutil>=5.9.0

# Optional models (may not be available)
catboost>=1.2.0
pytorch-tabnet>=4.0.0

# Telegram (optional)
telethon>=1.30.0

# Development
pytest>=7.0.0
